Reformat the layout for the project and rename some classes to make 2 signalr projects consistent (#333)
