+++ src/Microsoft.Azure.SignalR.AspNet/ClientConnections/ClientConnectionManager.cs
