+++ src/Microsoft.Azure.SignalR.AspNet/ClientConnections/IClientConnectionManager.cs
