+++ src/Microsoft.Azure.SignalR.AspNet/HubHost/AppMessage.cs
