+++ src/Microsoft.Azure.SignalR.AspNet/HubHost/EmptyProtectedData.cs
