+++ src/Microsoft.Azure.SignalR.AspNet/HubHost/HubMessage.cs
