+++ src/Microsoft.Azure.SignalR.AspNet/HubHost/IMessageParser.cs
