+++ src/Microsoft.Azure.SignalR.AspNet/HubHost/IServiceTransport.cs
