+++ src/Microsoft.Azure.SignalR.AspNet/HubHost/SignalRMessageParser.cs
