+++ src/Microsoft.Azure.SignalR.AspNet/ServerConnections/IServiceConnectionManager.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
 
 namespace Microsoft.Azure.SignalR.AspNet
 {
