+++ src/Microsoft.Azure.SignalR.AspNet/ServerConnections/ServiceConnection.Log.cs
