+++ src/Microsoft.Azure.SignalR.AspNet/ServerConnections/ServiceConnection.cs
