+++ src/Microsoft.Azure.SignalR.AspNet/ServerConnections/ServiceConnectionExtensions.cs
