+++ src/Microsoft.Azure.SignalR.AspNet/ServerConnections/ServiceConnectionManager.cs
