+++ src/Microsoft.Azure.SignalR/ClientConnections/ClientConnectionFactory.cs
