+++ src/Microsoft.Azure.SignalR/ClientConnections/ClientConnectionManager.cs
