+++ src/Microsoft.Azure.SignalR/ClientConnections/IClientConnectionFactory.cs
