+++ src/Microsoft.Azure.SignalR/ClientConnections/IClientConnectionManager.cs
