+++ src/Microsoft.Azure.SignalR/EndpointProvider/ServiceEndpointManager.cs
