+++ src/Microsoft.Azure.SignalR/Internals/AzureSignalRMarkerService.cs
