+++ src/Microsoft.Azure.SignalR/ServerConnections/IServiceConnectionManager.cs
