+++ src/Microsoft.Azure.SignalR/ServerConnections/ServiceConnection.Log.cs
