+++ src/Microsoft.Azure.SignalR/ServerConnections/ServiceConnection.cs
