+++ src/Microsoft.Azure.SignalR/ServerConnections/ServiceConnectionContext.cs
