+++ src/Microsoft.Azure.SignalR/ServerConnections/ServiceConnectionManager.cs
