+++ src/Microsoft.Azure.SignalR/ServiceOptionsSetup.cs
